+++ db/migrate/20241030212657_remove_fields_from_availability_tutors.rb
@@ -1,0 +1,7 @@
+class RemoveFieldsFromAvailabilityTutors < ActiveRecord::Migration[7.2]
+  def change
+    remove_column :availability_tutors, :description, :string
+    remove_column :availability_tutors, :link, :string
+    remove_column :availability_tutors, :availability, :string
+  end
+end