+++ spec/requests/topic_spec.rb
@@ -6,142 +6,200 @@ RSpec.describe "Topics", type: :request do
   let!(:subject) { FactoryBot.create(:subject, university: university) }
   let!(:topic) { FactoryBot.create(:topic, subject: subject) }
   let!(:token) { JsonWebTokenService.encode(user_id: user.id) }
+  let!(:availability_tutor) { FactoryBot.create(:availability_tutor, user: user, topic: topic) }
 
   describe "GET /topics" do
-    let!(:availability_tutor) { FactoryBot.create(:availability_tutor, user: user, topic: topic) }
     context "Without params" do
       it "returns http success" do
         get "/topics",
         headers: { 'Authorization': "Bearer #{token}" }
         expect(response).to have_http_status(:success)
+        expect(JSON.parse(response.body).size).to eq(1)
+      end
+
+      it "returns empty array" do
+        Topic.destroy_all
+        get "/topics",
+        headers: { 'Authorization': "Bearer #{token}" }
+        expect(response).to have_http_status(:success)
+        expect(JSON.parse(response.body)).to eq([])
       end
     end
 
     context "With user_id" do
-      let!(:user1) { FactoryBot.create(:user) }
-      let!(:availability_tutor1) { FactoryBot.create(:availability_tutor, user: user1, topic: topic) }
-
       it "returns http success with topics given by the tutor" do
         get "/topics", params: {
           'user_id': user.id
         },
         headers: { 'Authorization': "Bearer #{token}" }
         expect(response).to have_http_status(:success)
-        expect(response.body).to include(user.id)
-        expect(response.body).to_not include(user1.id)
+        parsed_response = JSON.parse(response.body)
+        expect(parsed_response.size).to eq(1)
+        topic = Topic.find(parsed_response.first["id"])
+        expect(topic.tutor.id).to eq(user.id)
       end
     end
 
     context "With subject_id" do
-      let!(:user1) { FactoryBot.create(:user) }
-      let!(:subject1) { FactoryBot.create(:subject, university: university) }
-      let!(:topic1) { FactoryBot.create(:topic, subject: subject1) }
-      let!(:availability_tutor1) { FactoryBot.create(:availability_tutor, user: user1, topic: topic1) }
-
       it "returns http success with topics belonging to any subject" do
         get "/topics", params: {
           'subject_id': subject.id
         },
         headers: { 'Authorization': "Bearer #{token}" }
         expect(response).to have_http_status(:success)
-        expect(response.body).to include(subject.id.to_s)
-        expect(response.body).to_not include(subject1.id.to_s)
+        parsed_response = JSON.parse(response.body)
+        expect(parsed_response.size).to eq(1)
+        topic = Topic.find(parsed_response.first["id"])
+        expect(topic.subject.id).to eq(subject.id)
+      end
+
+      it "returns empty array" do
+        get "/topics", params: {
+          'subject_id': 1
+        },
+        headers: { 'Authorization': "Bearer #{token}" }
+        expect(response).to have_http_status(:success)
+        expect(JSON.parse(response.body)).to eq([])
+      end
+    end
+
+    context "With subject_id and user_id" do
+      let!(:user2) { FactoryBot.create(:user) }
+      let!(:subject2) { FactoryBot.create(:subject, university: university) }
+      let!(:topic2) { FactoryBot.create(:topic, subject: subject2) }
+      let!(:availability_tutor2) { FactoryBot.create(:availability_tutor, user: user2, topic: topic2) }
+      it "returns http success" do
+        get "/topics", params: {
+          'subject_id': subject2.id,
+          'user_id': user2.id
+        },
+        headers: { 'Authorization': "Bearer #{token}" }
+        expect(response).to have_http_status(:success)
+        parsed_response = JSON.parse(response.body)
+        expect(parsed_response.size).to eq(1)
+        topic = Topic.find(parsed_response.first["id"])
+        expect(topic.subject.id).to eq(subject2.id)
+        expect(topic.tutor.id).to eq(user2.id)
+      end
+
+      it "returns empty array" do
+        get "/topics", params: {
+          'subject_id': subject2.id,
+          'user_id': user.id
+        },
+        headers: { 'Authorization': "Bearer #{token}" }
+        expect(response).to have_http_status(:success)
+        expect(JSON.parse(response.body)).to eq([])
       end
     end
   end
 
   describe "GET /topics/:id" do
-    let!(:availability_tutor) { FactoryBot.create(:availability_tutor, user: user, topic: topic) }
-    it "returns http not found" do
-      get "/topics/1",
-      headers: { 'Authorization': "Bearer #{token}" }
-      expect(response.body).to include("Topic not found")
-    end
-
-    it "returns http success without interesteds" do
-      get "/topics/#{availability_tutor.id}",
-      headers: { 'Authorization': "Bearer #{token}" }
-      expect(response).to have_http_status(:ok)
-      parsed = JSON.parse(response.body)
-      expect(parsed["availability_id"]).to eq(availability_tutor.id)
-      expect(parsed["interested_users"]).to eq([])
-    end
-
-    it "returns http success with interesteds" do
-      user_interested = FactoryBot.create(:user)
-      FactoryBot.create(:interested, user: user_interested, availability_tutor: availability_tutor)
-      get "/topics/#{availability_tutor.id}",
-      headers: { 'Authorization': "Bearer #{token}" }
-      expect(response).to have_http_status(:ok)
-      parsed = JSON.parse(response.body)
-      expect(parsed["availability_id"]).to eq(availability_tutor.id)
-      expect(parsed["interesteds"]).to eq(1)
-      expect(parsed["interested_users"].first["id"]).to eq(user_interested.id)
-    end
+    # to do
   end
 
-  describe "GET /proposed_topics" do
-    context "Without proposed topics" do
-      it "returns http success" do
-        get "/proposed_topics",
+  describe "POST /topics" do
+    context "With valid params" do
+      it "returns http created" do
+        post "/topics", params: {
+          topic: {
+            name: "New topic",
+            description: "New description",
+            link: "New link",
+            show_email: true,
+            subject_id: subject.id
+          }
+        },
         headers: { 'Authorization': "Bearer #{token}" }
-        expect(response).to have_http_status(:success)
-        expect(response.body).to eq("[]")
+        expect(response).to have_http_status(:created)
+        parsed = JSON.parse(response.body)
+        expect(parsed["message"]).to eq(I18n.t("success.topics.created"))
+        newTopic = Topic.find_by(name: "New topic")
+        expect(newTopic).not_to be_nil
+        expect(newTopic.description).to eq("New description")
+        expect(newTopic.link).to eq("New link")
+        expect(newTopic.show_email).to eq(true)
+        expect(newTopic.subject.id).to eq(subject.id)
       end
     end
 
-    context "With proposed topics" do
-      it "returns http success" do
-        get "/proposed_topics",
+    context "With invalid params" do
+      it "subject_id not found" do
+        post "/topics", params: {
+          topic: {
+            name: "New topic",
+            description: "New description",
+            link: "New link",
+            show_email: true,
+            subject_id: 0
+          }
+        },
         headers: { 'Authorization': "Bearer #{token}" }
-        expect(response).to have_http_status(:success)
-        expect(response).to_not be_nil
+        expect(response).to have_http_status(:unprocessable_entity)
+        parsed = JSON.parse(response.body)
+        expect(parsed["error"]).to eq("Validation failed: Subject must exist")
+      end
+
+      it "no subject_id provided" do
+        post "/topics", params: {
+          topic: {
+            name: "New topic",
+            description: "New description",
+            link: "New link",
+            show_email: true,
+            subject_id: nil
+          }
+        },
+        headers: { 'Authorization': "Bearer #{token}" }
+        expect(response).to have_http_status(:unprocessable_entity)
+        expect(JSON.parse(response.body)["error"]).to eq("Validation failed: Subject must exist, Subject can't be blank")
+      end
+
+      it "no name provided" do
+        post "/topics", params: {
+          topic: {
+            name: nil,
+            description: "New description",
+            link: "New link",
+            show_email: true,
+            subject_id: subject.id
+          }
+        },
+        headers: { 'Authorization': "Bearer #{token}" }
+        expect(response).to have_http_status(:unprocessable_entity)
+        expect(JSON.parse(response.body)["error"]).to eq("Validation failed: Name can't be blank")
       end
     end
-  end
 
-  describe "DELETE /proposed_topics/:availability_id" do
-    let!(:availability_tutor) { FactoryBot.create(:availability_tutor, user: user, topic: topic) }
-
-    context "Without meets created" do
-      it "returns http success without interesteds" do
-        delete "/proposed_topics/#{availability_tutor.id}",
-        headers: { 'Authorization': "Bearer #{token}" }
-        expect(response).to have_http_status(:success)
-        expect(Topic.find_by(id: topic.id)).to be_nil
-        expect(AvailabilityTutor.find_by(id: availability_tutor.id)).to be_nil
-      end
-
-      it "returns http success with interesteds" do
-        user_interested = FactoryBot.create(:user)
-        FactoryBot.create(:interested, user: user_interested,  availability_tutor: availability_tutor)
-        delete "/proposed_topics/#{availability_tutor.id}",
-        headers: { 'Authorization': "Bearer #{token}" }
-        expect(response).to have_http_status(:success)
-        expect(Interested.find_by(user_id: user_interested.id)).to be_nil
-      end
+    it "invalid show_email" do
+      post "/topics", params: {
+        topic: {
+          name: "raaaaaa",
+          description: "New description",
+          link: "New link",
+          show_email: "invalid",
+          subject_id: subject.id
+        }
+      },
+      headers: { 'Authorization': "Bearer #{token}" }
+      expect(response).to have_http_status(:created)
+      expect(JSON.parse(response.body)["message"]).to eq(I18n.t("success.topics.created"))
+      expect(Topic.find_by(name: "raaaaaa").show_email).to eq(true)
     end
 
-    context "With meets created" do
-      let!(:meet) { FactoryBot.create(:meet, availability_tutor: availability_tutor) }
-
-      it "returns http success without intrested in meet" do
-        delete "/proposed_topics/#{availability_tutor.id}",
-        headers: { 'Authorization': "Bearer #{token}" }
-        expect(response).to have_http_status(:success)
-        expect(Meet.find_by(id: meet.id)).to be_nil
-      end
-
-      it "returns http success with intrested in meet" do
-        user_interested = FactoryBot.create(:user)
-        FactoryBot.create(:interested, user: user_interested, availability_tutor: availability_tutor)
-        FactoryBot.create(:participant, user: user_interested, meet: meet)
-        delete "/proposed_topics/#{availability_tutor.id}",
-        headers: { 'Authorization': "Bearer #{token}" }
-        expect(response).to have_http_status(:success)
-        expect(Meet.find_by(id: meet.id)).to be_nil
-        expect(Participant.find_by(user_id: user_interested.id)).to be_nil
-      end
+    it "show_email nil" do
+      post "/topics", params: {
+        topic: {
+          name: "raaaaaa",
+          description: "New description",
+          link: "New link",
+          show_email: nil,
+          subject_id: subject.id
+        }
+      },
+      headers: { 'Authorization': "Bearer #{token}" }
+      expect(response).to have_http_status(:unprocessable_entity)
+      expect(JSON.parse(response.body)["error"]).to eq("Validation failed: Show email is not included in the list")
     end
   end
 end